+++ app/controllers/wallaby/abstract_resources_controller.rb
@@ -2,6 +2,7 @@ module Wallaby
   # Generic CRUD controller
   class AbstractResourcesController < ::Wallaby::BaseController
     include ResourcesHelperMethods
+    include RailsOverridenMethods
     self.responder = ResourcesResponder
     respond_to :html
     respond_to :json
@@ -276,33 +277,6 @@ module Wallaby
     end
 
     protected
-
-    # Override origin ActionView::ViewPaths::ClassMethods#_prefixes
-    # to add more paths so that it could look up partials in the following
-    # order:
-    # - mounted_path/resources_name/action_name (e.g. `admin/products/index)
-    # - mounted_path/resources_name (e.g. `admin/products)
-    # - full_path_of_custom_resources_controller/action_name
-    #  (e.g. `management/custom_products/index)
-    # - full_path_of_custom_resources_controller
-    #  (e.g. `management/custom_products)
-    # - wallaby_resources_controller_name/action_name
-    #  (e.g. `wallaby/resources/index)
-    # - wallaby_resources_controller_name
-    #  (e.g. `wallaby/resources)
-    # @return [PrefixesBuilder]
-    def _prefixes
-      @_prefixes ||= PrefixesBuilder.new(
-        super, controller_path, current_resources_name, params
-      ).build
-    end
-
-    # A wrapped lookup content
-    # Its aim is to render string partial when given partial is not found
-    # @return [LookupContextWrapper]
-    def lookup_context
-      @_lookup_context ||= LookupContextWrapper.new super # rubocop:disable Naming/MemoizedInstanceVariableName, Metrics/LineLength
-    end
 
     # To paginate the collection but only when either `page` or `per` param is
     # given, or requesting HTML response
